Refactor logging system from LoggerFactory to Reporter pattern
- Replace packages/pino-adapter/ with packages/pino-log-reporter/ using Reporter pattern
- Consolidate logging implementation into packages/kori/src/logging/logger.ts
- Update API names: KoriReporter → KoriLogReporter, createKoriSimpleLoggerFactory → createKoriLoggerFactory
- Change logging method calls: req.log().child() → ctx.log().channel(), req.log() → ctx.log()
- Update all plugins and examples to use new logging API

BREAKING CHANGE: @korix/pino-adapter package removed, use @korix/pino-log-reporter instead
