--- packages/pino-adapter/src/index.ts
@@ -1,1 +1,0 @@
-export { createPinoKoriLoggerFactory } from './pino-logger-factory.js';
