+++ packages/pino-log-reporter/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@korix/pino-adapter",
+  "name": "@korix/pino-log-reporter",
   "version": "0.1.1",
-  "description": "Pino logger adapter for Kori framework",
+  "description": "Pino log reporter for Kori framework",
   "type": "module",
   "author": "mitz (@bufferings)",
   "license": "MIT",
