+++ packages/pino-log-reporter/src/index.ts
@@ -1,0 +1,1 @@
+export { createPinoLogReporter } from './pino-log-reporter.js';
