+++ packages/pino-log-reporter/tsconfig.json
